requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
decorator==5.0.9
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.7
django-allauth==0.45.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
ipython==7.27.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
oauthlib==3.1.1
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
PyJWT==1.7.1
pyparsing==2.4.7
pyspark==3.1.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sqlparse==0.4.1
traitlets==5.1.0
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
